added use strict to main code, separated simulator from ipcmainlisteners, added eslint and prop-types to dependencies
