+++ lib/file-tree.js
@@ -1,10 +1,12 @@
+'use strict'
 //Pass in path to root directory
-//returns object with all directories and files
-//sets up recursive watch on file tree
+//returns object representation with all directories and files
 //directory properties have property structure
 //path: STRING, files: ARRAY, opened: BOOL (default false), type:STRING
 //file properties have property structure
 //path:STRING, type:STRING
+
+//separate out the proj info stuff into another module? and then put info in local storage instead of a local file?
 
 const fs = require('fs');
 const path = require('path');
@@ -72,21 +74,22 @@ const getTree = (rootDirPath, callback) => {
               projInfo.htmlPath = filePath;
             }
 
-            //look for package.json and see if webpack is installed
+            //look for package.json and see if webpack is installed and react-hot-loader is installed
             else if (!projInfo.webpack && file.search(/package.json/) !== -1 && filePath === path.join(projInfo.rootPath, file)) {
               fs.readFile(filePath, { encoding: 'utf-8' }, (err, data) => {
                 if (data.search(/webpack/) !== -1) projInfo.webpack = true;
+                if (data.search(/react-hot-loader/) !== -1) projInfo.hotLoad = true;
               })
             } 
 
-            //check index.js for AppContainer, which means there is hot loading. not very robust.
-            else if (!projInfo.hotLoad && filePath.search(/.*node\_modules.*/) === -1 && file.search(/index.js/)) {
-              fs.readFile(filePath, { encoding: 'utf-8' }, (err, data) => {
-                if (data.search(/react-hot-loader/) !== -1) {
-                  projInfo.hotLoad = true;
-                }
-              })
-            }
+            
+            // else if (!projInfo.hotLoad && filePath.search(/.*node\_modules.*/) === -1 && file.search(/index.js/)) {
+            //   fs.readFile(filePath, { encoding: 'utf-8' }, (err, data) => {
+            //     if (data.search(/react-hot-loader/) !== -1) {
+            //       projInfo.hotLoad = true;
+            //     }
+            //   })
+            // }
 
             if (!--pending) {
               callback(fileTree);
