+++ lib/item-schema.js
@@ -1,3 +1,5 @@
+'use strict'
+
 let uniqueId = 0;
 
 function File(path, name) {
