+++ package.json
@@ -18,6 +18,7 @@
     "fs": "0.0.1-security",
     "monaco-editor": "^0.8.3",
     "path": "^0.12.7",
+    "prop-types": "^15.5.8",
     "react": "^15.4.2",
     "react-dom": "^15.4.2",
     "react-redux": "^5.0.3",
@@ -39,6 +40,7 @@
     "electron-packager": "^8.5.2",
     "electron-reload": "^1.1.0",
     "enzyme": "^2.8.0",
+    "eslint": "^3.19.0",
     "expect": "^1.20.2",
     "mocha": "^3.2.0",
     "react-addons-test-utils": "^15.4.2",
